refactor: split synthesizer surface out of core MCP path
- Move lm-studio-synthesizer.ts and ai-trace-log.ts into src/runtime/synthesizer/
- Create narrow barrel export (index.ts) so runtime.ts only imports createSynthesizerFromEnv
- Extract lm-check CLI subcommand into standalone src/runtime/synthesizer/lm-check.ts
- Core CLI now only carries FPF-indexing subcommands (refresh, status, query, inspect, read-doc, trace)
- Split tool-contracts.ts into public-contracts.ts (3 public tools) and expert-contracts.ts (6 expert tools)
- Delete monolithic tool-contracts.ts
- Update README and PR template to reference new contract file structure

Closes #37
