+++ src/mcp/expert-contracts.ts
@@ -1,15 +1,13 @@
 import { z } from 'zod';
 
-export const answerModeSchema = z.enum(['compact', 'verbose', 'proof']);
+import {
+  answerModeSchema,
+  answerStatusSchema,
+  snapshotWithRebuildSchema,
+} from './public-contracts.js';
+
 export const nodeKindSchema = z.enum(['pattern', 'route', 'lexeme']);
 export const selectorKindSchema = z.enum(['auto', 'id', 'route', 'lexeme']);
-export const answerStatusSchema = z.enum([
-  'ok',
-  'not_found',
-  'ambiguous',
-  'unsupported',
-  'stale_snapshot_prevented',
-]);
 export const anchorRoleSchema = z.enum([
   'definition',
   'solution',
@@ -26,14 +24,6 @@ export const buildReasonSchema = z.enum([
   'source_hash_changed',
   'snapshot_current',
 ]);
-export const observabilityFormatSchema = z.enum(['flat', 'tree', 'normalized']);
-export const observabilityLogLevelSchema = z.enum([
-  'debug',
-  'info',
-  'warn',
-  'error',
-  'fatal',
-]);
 export const resolvedAsSchema = z.enum(['id', 'route', 'lexeme', 'not_found']);
 export const inspectStatusSchema = z.enum(['ok', 'not_found']);
 export const frontierOriginSchema = z.enum([
@@ -45,15 +35,6 @@ export const frontierOriginSchema = z.enum([
   'session_context',
 ]);
 export const expandedCitationStatusSchema = z.enum(['ok', 'not_found']);
-export const lmStudioApiStyleSchema = z.enum(['responses', 'lmstudio_chat', 'chat_completions']);
-
-export const relationEdgeSchema = z
-  .object({
-    from: z.string(),
-    relation: z.string(),
-    to: z.string(),
-  })
-  .strict();
 
 export const inspectNeighborSchema = z
   .object({
@@ -110,14 +91,6 @@ export const compiledNodeSchema = z
   })
   .strict();
 
-export const snapshotWithRebuildSchema = z
-  .object({
-    sourceHash: z.string(),
-    builtAt: z.string(),
-    rebuilt: z.boolean(),
-  })
-  .strict();
-
 export const buildAuditSchema = z
   .object({
     sourcePath: z.string(),
@@ -151,78 +124,6 @@ export const buildAuditSchema = z
       })
       .strict(),
     artifacts: z.record(z.string(), z.string()),
-  })
-  .strict();
-
-export const queryResultSchema = z
-  .object({
-    mode: answerModeSchema,
-    question: z.string(),
-    answer: z.string(),
-    ids: z.array(z.string()),
-    relations: z.array(relationEdgeSchema),
-    constraints: z.array(z.string()),
-    citations: z.array(z.string()),
-    confidence: z.number(),
-    gaps: z.array(z.string()),
-    snapshot: snapshotWithRebuildSchema,
-    status: answerStatusSchema,
-    groundingChain: z.array(z.string()).optional(),
-  })
-  .strict();
-
-export const askFpfResultSchema = z
-  .object({
-    question: z.string(),
-    mode: answerModeSchema,
-    markdown: z.string(),
-    ids: z.array(z.string()),
-    citations: z.array(z.string()),
-    constraints: z.array(z.string()),
-    gaps: z.array(z.string()),
-    confidence: z.number(),
-    status: answerStatusSchema,
-    snapshot: snapshotWithRebuildSchema,
-    groundingChain: z.array(z.string()).optional(),
-  })
-  .strict();
-
-export const runtimeStatusSchema = z
-  .object({
-    sourcePath: z.string(),
-    sourceHash: z.string().optional(),
-    builtAt: z.string().optional(),
-    snapshotExists: z.boolean(),
-    currentSourceHash: z.string(),
-    fresh: z.boolean(),
-    compilerMode: z.literal('local_vectorless'),
-    artifacts: z.record(z.string(), z.boolean()),
-    synthesizer: z
-      .object({
-        configured: z.boolean(),
-        provider: z.string().optional(),
-        model: z.string().optional(),
-        baseUrl: z.string().optional(),
-        apiStyle: lmStudioApiStyleSchema.optional(),
-      })
-      .strict(),
-    observability: z
-      .object({
-        configured: z.boolean(),
-        filePath: z.string(),
-        format: observabilityFormatSchema,
-        includeInternalSpans: z.boolean(),
-        logLevel: observabilityLogLevelSchema,
-        excludeModelChunks: z.boolean(),
-      })
-      .strict(),
-    sessionCache: z
-      .object({
-        enabled: z.boolean(),
-        maxSessions: z.number(),
-        activeSessions: z.number(),
-      })
-      .strict(),
   })
   .strict();
 
@@ -387,7 +288,7 @@ export const refreshFpfIndexInputSchema = z
   })
   .strict();
 
-export const queryFpfSpecInputSchema = z
+export const traceFpfPathInputSchema = z
   .object({
     question: z.string().min(1),
     mode: answerModeSchema.optional(),
@@ -395,17 +296,6 @@ export const queryFpfSpecInputSchema = z
     sessionId: z.string().min(1).optional(),
   })
   .strict();
-
-export const askFpfInputSchema = z
-  .object({
-    question: z.string().min(1),
-    mode: answerModeSchema.optional(),
-    forceRefresh: z.boolean().optional(),
-    sessionId: z.string().min(1).optional(),
-  })
-  .strict();
-
-export const getFpfIndexStatusInputSchema = z.object({}).strict();
 
 export const inspectFpfNodeInputSchema = z
   .object({
@@ -437,19 +327,7 @@ export const expandFpfCitationsInputSchema = z
   })
   .strict();
 
-export const traceFpfPathInputSchema = z
-  .object({
-    question: z.string().min(1),
-    mode: answerModeSchema.optional(),
-    forceRefresh: z.boolean().optional(),
-    sessionId: z.string().min(1).optional(),
-  })
-  .strict();
-
 export type RefreshFpfIndexInput = z.infer<typeof refreshFpfIndexInputSchema>;
-export type QueryFpfSpecInput = z.infer<typeof queryFpfSpecInputSchema>;
-export type AskFpfInput = z.infer<typeof askFpfInputSchema>;
-export type GetFpfIndexStatusInput = z.infer<typeof getFpfIndexStatusInputSchema>;
 export type InspectFpfNodeInput = z.infer<typeof inspectFpfNodeInputSchema>;
 export type ReadFpfDocInput = z.infer<typeof readFpfDocInputSchema>;
 export type InspectFpfAnchorInput = z.infer<typeof inspectFpfAnchorInputSchema>;
