+++ src/runtime/synthesizer/ai-trace-log.ts
@@ -1,8 +1,8 @@
 import { randomUUID } from 'node:crypto';
 import { appendFile } from 'node:fs/promises';
 
-import { resolveLogPath } from '../logging/file-paths.js';
-import type { AnswerMode, TraceResult } from './types.js';
+import { resolveLogPath } from '../../logging/file-paths.js';
+import type { AnswerMode, TraceResult } from '../types.js';
 import type { LmStudioApiStyle } from './lm-studio-synthesizer.js';
 
 export interface AiTraceRequestLog {
