+++ src/runtime/synthesizer/index.ts
@@ -1,0 +1,1 @@
+export { createSynthesizerFromEnv } from './lm-studio-synthesizer.js';
